--- a/fd/file.d.go
+++ b/fd/file.d.go
@@ -222,6 +222,11 @@ func (f *FileD) setupOutput(p *pipeline.Pipeline, pipelineConfig *cfg.PipelineCo
 		PluginRuntimeInfo: f.instantiatePlugin(info),
 	})
 
+	p.SetDeadQueueOutput(&pipeline.OutputPluginInfo{
+		PluginStaticInfo:  info.DeadQueueInfo,
+		PluginRuntimeInfo: f.instantiatePlugin(info.DeadQueueInfo),
+	})
+
 	return nil
 }
 
@@ -249,6 +254,19 @@ func (f *FileD) getStaticInfo(pipelineConfig *cfg.PipelineConfig, pluginKind pip
 	if err != nil {
 		return nil, err
 	}
+	deadqueue := configJSON.Get("deadqueue")
+	var deadqueueInfo *pipeline.PluginStaticInfo
+	if deadqueue.MustMap() != nil {
+		deadqueueType := deadqueue.Get("type").MustString()
+		if deadqueueType == "" {
+			return nil, fmt.Errorf("%s doesn't have type", pluginKind)
+		}
+		deadqueueInfo, err = f.plugins.Get(pluginKind, deadqueueType)
+		if err != nil {
+			return nil, err
+		}
+		configJSON.Del("deadqueue")
+	}
 	configJson, err := configJSON.Encode()
 	if err != nil {
 		logger.Panicf("can't create config json for %s", t)
@@ -260,6 +278,9 @@ func (f *FileD) getStaticInfo(pipelineConfig *cfg.PipelineConfig, pluginKind pip
 
 	infoCopy := *info
 	infoCopy.Config = config
+	if deadqueueInfo != nil {
+		infoCopy.DeadQueueInfo = deadqueueInfo
+	}
 
 	return &infoCopy, nil
 }