--- a/pipeline/plugin.go
+++ b/pipeline/plugin.go
@@ -78,7 +78,8 @@ type PluginStaticInfo struct {
 	// Every plugin can provide their own API through Endpoints.
 	Endpoints         map[string]func(http.ResponseWriter, *http.Request)
 	AdditionalActions []string // used only for input plugins, defines actions that should be run right after input plugin with input config
-	DeadQueueInfo     *PluginStaticInfo
+	// TODO: maybe to OutputPluginStaticInfo cause uses by output and action plugins?
+	DeadQueueInfo *PluginStaticInfo
 }
 
 type PluginRuntimeInfo struct {