--- a/pipeline/plugin.go
+++ b/pipeline/plugin.go
@@ -59,6 +59,7 @@ type ActionPluginParams struct {
 type OutputPluginParams struct {
 	PluginDefaultParams
 	Controller OutputPluginController
+	Router     Router
 	Logger     *zap.SugaredLogger
 }