--- a/plugin/output/clickhouse/clickhouse.go
+++ b/plugin/output/clickhouse/clickhouse.go
@@ -58,6 +58,8 @@ type Plugin struct {
 	// plugin metrics
 	insertErrorsMetric prometheus.Counter
 	queriesCountMetric prometheus.Counter
+
+	router pipeline.Router
 }
 
 type Setting struct {
@@ -437,15 +439,17 @@ func (p *Plugin) Start(config pipeline.AnyConfig, params *pipeline.OutputPluginP
 		MetricCtl:      params.MetricCtl,
 	}
 
+	p.router = params.Router
 	backoffOpts := pipeline.BackoffOpts{
-		MinRetention: p.config.Retention_,
-		Multiplier:   float64(p.config.RetentionExponentMultiplier),
-		AttemptNum:   p.config.Retry,
+		MinRetention:         p.config.Retention_,
+		Multiplier:           float64(p.config.RetentionExponentMultiplier),
+		AttemptNum:           p.config.Retry,
+		DeadQueueIsAvailable: p.router.DeadQueueIsAvailable(),
 	}
 
-	onError := func(err error, _ []*pipeline.Event) {
+	onError := func(err error, events []*pipeline.Event) {
 		var level zapcore.Level
-		if p.config.FatalOnFailedInsert {
+		if p.config.FatalOnFailedInsert && !p.router.DeadQueueIsAvailable() {
 			level = zapcore.FatalLevel
 		} else {
 			level = zapcore.ErrorLevel
@@ -454,6 +458,10 @@ func (p *Plugin) Start(config pipeline.AnyConfig, params *pipeline.OutputPluginP
 		p.logger.Log(level, "can't insert to the table", zap.Error(err),
 			zap.Int("retries", p.config.Retry),
 			zap.String("table", p.config.Table))
+
+		for i := range events {
+			p.router.Fail(events[i])
+		}
 	}
 
 	p.batcher = pipeline.NewRetriableBatcher(