--- a/plugin/output/clickhouse/clickhouse.go
+++ b/plugin/output/clickhouse/clickhouse.go
@@ -59,7 +59,7 @@ type Plugin struct {
 	insertErrorsMetric prometheus.Counter
 	queriesCountMetric prometheus.Counter
 
-	router pipeline.Router
+	router *pipeline.Router
 }
 
 type Setting struct {
@@ -444,12 +444,12 @@ func (p *Plugin) Start(config pipeline.AnyConfig, params *pipeline.OutputPluginP
 		MinRetention:         p.config.Retention_,
 		Multiplier:           float64(p.config.RetentionExponentMultiplier),
 		AttemptNum:           p.config.Retry,
-		DeadQueueIsAvailable: p.router.DeadQueueIsAvailable(),
+		IsDeadQueueAvailable: p.router.IsDeadQueueAvailable(),
 	}
 
 	onError := func(err error, events []*pipeline.Event) {
 		var level zapcore.Level
-		if p.config.FatalOnFailedInsert && !p.router.DeadQueueIsAvailable() {
+		if p.config.FatalOnFailedInsert && !p.router.IsDeadQueueAvailable() {
 			level = zapcore.FatalLevel
 		} else {
 			level = zapcore.ErrorLevel