--- a/plugin/output/stdout/stdout.go
+++ b/plugin/output/stdout/stdout.go
@@ -40,5 +40,4 @@ func (p *Plugin) Out(event *pipeline.Event) {
 	// nolint: forbidigo
 	fmt.Println(event.Root.EncodeToString())
 	p.controller.Commit(event)
-	p.router.Ack(event)
 }