# Description

When the main output fails, it goes into a dead queue

## Examples

### Dead queue to the reserve elasticsearch

```yaml
main_pipeline:
  input:
    type: kafka
    brokers:
      - kafka:9092
    topics:
      - logs
  output:
    type: elasticsearch
    workers_count: 32
    endpoints:
      - http://elasticsearch-1:9200
    # route to reserve elasticsearch
    deadqueue:
      endpoints:
        - http://elasticsearch-2:9200
      type: elasticsearch
```

### Dead queue with second kafka topic and low priority consumer

```yaml
main_pipeline:
  input:
    type: kafka
    brokers:
      - kafka:9092
    topics:
      - logs
  output:
    type: elasticsearch
    workers_count: 32
    endpoints:
      - http://elasticsearch:9200
    # route to deadqueue pipeline
    deadqueue:
      brokers:
      - kafka:9092
      default_topic: logs-deadqueue
      type: kafka

deadqueue_pipeline:
  input:
    type: kafka
    brokers:
      - kafka:9092
    topics:
      - logs-deadqueue
  output:
    type: elasticsearch
    workers_count: 1 # low priority
    fatal_on_failed_insert: false
    endpoints:
      - http://elasticsearch:9200
```