@DmitryRomanov 2025-08-08
Benches:

```
# before
BenchmarkLightJsonReadPar-8            1        3454751723 ns/op         420.07 MB/s    191420944 B/op   2129176 allocs/op

# after
BenchmarkLightJsonReadPar-8            1        3350755643 ns/op         433.11 MB/s    229879008 B/op   2557434 allocs/op
```


```
# Before

go mod download && go run ./bench/bench.go ./bench/logger.go
info    started, waiting for first request
info    stats bytes=280Mb, requests=377, throughput: 280.86Mb/s
info    stats bytes=482Mb, requests=642, throughput: 242.26Mb/s
info    stats bytes=678Mb, requests=910, throughput: 226.87Mb/s
info    stats bytes=914Mb, requests=1226, throughput: 228.77Mb/s
info    stats bytes=1170Mb, requests=1574, throughput: 234.73Mb/s
info    stats bytes=1399Mb, requests=1880, throughput: 234.26Mb/s
info    stats bytes=1529Mb, requests=2055, throughput: 218.09Mb/s
info    stats bytes=1703Mb, requests=2289, throughput: 212.85Mb/s
info    stats bytes=1910Mb, requests=2565, throughput: 212.05Mb/s
info    final result: 212.93Mb/s


# After

go mod download && go run ./bench/bench.go ./bench/logger.go
info    started, waiting for first request
info    stats bytes=274Mb, requests=368, throughput: 277.27Mb/s
info    stats bytes=562Mb, requests=753, throughput: 282.59Mb/s
info    stats bytes=872Mb, requests=1169, throughput: 291.11Mb/s
info    stats bytes=1166Mb, requests=1568, throughput: 292.08Mb/s
info    stats bytes=1384Mb, requests=1862, throughput: 276.96Mb/s
info    stats bytes=1676Mb, requests=2253, throughput: 279.60Mb/s
info    stats bytes=1951Mb, requests=2619, throughput: 279.23Mb/s
info    stats bytes=2193Mb, requests=2948, throughput: 274.32Mb/s
info    stats bytes=2398Mb, requests=3222, throughput: 266.61Mb/s
info    final result: 261.65Mb/s
```